Add solution for CCC 2020 J5/S2 - Escape Room. This is not completed that it TLE on the last unit test. Algo needs to be improved later.
Todo: improve the algo to pass Unit test batch 7.

Implemented a BFS-based solution to determine if the bottom-right cell of the grid can be reached based on the problem's constraints. The algorithm uses a queue to traverse valid moves while marking visited cells and handles edge cases for m and n.